added lookup builder
